impl `Future` for external `Either` rather than defining our own
